backend/routes/progress.py
from flask import Blueprint, jsonify
from flask_jwt_extended import jwt_required, get_jwt_identity
from models import Progress

progress_bp = Blueprint('progress', __name__)

@progress_bp.route('/', methods=['GET'])
@jwt_required()
def get_progress():
    user_id = get_jwt_identity()
    progress = Progress.query.filter_by(user_id=user_id).all()
    return jsonify({'history': [
        {'date': p.date.isoformat(), 'score': p.score}
        for p in progress
    ]}) 